Fix handling of one-off sessions
If a control socket and a dead pidfile are found for a session that isn't in our config file, remove both files and do not add return it as a valid session option. If a control socket is found, but no pidfile is found, remove the control socket automatically. Regardless of the socket being connected to a running process, there's no way to send signals to the process.