Logging revamp / cleanup
Logging now correctly supports STDOUT, STDERR and syslog, depending on configuration options.

By default, select messages are logged to stdout.

Add -v (or set your log_level to LOG_INFO) to see a few additional messages on STDERR and if a facility is defined, in syslog.

Add another -v (or set your log_level to LOG_DEBUG) to see full internal tracing of events, to pinpoint a specific issue. This is logged to both STDERR and to a syslog facility, if defined.

Finally, only unset the noclean flag once we've passed the check to make sure this host/session handler isn't already running. This simplifies the exit hander greatly.