Support refresh-client
With the release of tmux 3.0 (how did I possibly miss this release?),
tmux now has the no-output flag that can be set on a control mode
client. When set, %output lines are no longer sent over the control
session. This is a nice reduction in network traffic.

Support was added to get the tmux version and store it in the config
hash. A new command / handler called RefreshClient has been added that
returns the client_name, session_group, and session_name. If the tmux
major version is 3 or higher, refresh-client is called to disable output
on the control mode client.

RefreshClient:client-19190,global-session,auri-global-session-control

tmux refresh-client -t client-19190 -F no-output