More discoverabiility
Add --print|o

Use this option to print all configured options for a given
host/session. This lets you easily see which options were pulled from
defaults, which were set on the CLI, and which were used from a
configuration file.

Add the --known|k --known|k=,

Use this option to print all known tmuxc instances. It will print
anything found in the pid/control directories, as well as anything
configured in the tmuxc configuration file. The PID and control socket
path are printed if the instance is running. When invoked via -k=,
CSV output is enabled, allowing easy integration with scripts. Any
delimiter can be used, but in most cases it will either need to be
quoted or escaped.