Add one-off session launcher
This adds the ability to launch a one-off session via tmuxc -S . You can
input a session in the following format:

- session
- session@localhost
- session@hostname

The first two formats are equivalent to each other, and will create a
new tmux session on the local machine. These sessions are ephemeral -
once the controlling tmuxc instance is destroyed, they are forgotten.