Further command line length reduction
Switch to attach, instead of attach-session

Move the select-window operation to operate in the same tmux bin call
that creates our cloned session for a window.

Instead of setting the destroy-unattached option after we select a
window and attach to it, inject this directly into our control mode
client AFTER a client has attached to a session. If it's done before a
client is attached to a session, the session is destroyed.

Add examples from a control mode session for the other events we
process.

Update the text around the events we do process, to make the external
conditions that produce that event a little less ambiguous.

Split off the major version and compare that - since we can't cast 3.0a
to an integeger.