SSH Speedup, part 1
Instead of running multiple commands over an SSH tunnel, we build a
shell script that is then executed on the remote host (or local, if
operating on localhost). The script is written in POSIX sh, so it
should operate on any host capable of running tmux.

From startup to first terminal on a remote host, we have the following timing:

+ 300ms to detect if a ControlMaster SSH session is running, start it, and
then create our setup script

+ 300ms to check for / create our session, our control session, and return
the tmux version

+ 500ms to clone our base session, attach to it over SSH and spawn a
terminal.
---
= 1100ms from tmuxc start to having a local terminal running with a shell
on the remote host.

With the ControlMaster SSH session already running, a second startup
looks like this:

+ 1ms to check if our ControlMaster SSH session is running

+ 100ms to check for our session, control session and return the tmux
version

+ 500ms to clone our base session, attach to it over SSH and spawn a
terminal
---
= 601ms from tmuxc start to having a local terminal running with a
shell on the remote host.